# dev/core#5237: improve caching of afform metadata