--- a/ext/afform/core/CRM/Afform/AfformScanner.php
+++ b/ext/afform/core/CRM/Afform/AfformScanner.php
@@ -140,8 +140,12 @@ class CRM_Afform_AfformScanner {
    * @see \Civi\Api4\Afform::getFields
    */
   public function getMeta(string $name, bool $getLayout = FALSE): ?array {
+    $cacheKey = 'afform.meta.' . $name;
+    if ($getLayout) {
+      $cacheKey .= '.getLayout';
+    }
     if ($this->isUseCachedPaths()) {
-      $cachedDefn = $this->cache->get('afform.meta.' . $name);
+      $cachedDefn = $this->cache->get($cacheKey);
       if ($cachedDefn !== NULL) {
         return $cachedDefn;
       }
@@ -180,7 +184,7 @@ class CRM_Afform_AfformScanner {
     $defn['name'] = $name;
     $defn['modified_date'] = date('Y-m-d H:i:s', $mtime);
     if ($this->isUseCachedPaths()) {
-      $this->cache->set('afform.meta.' . $name, $defn);
+      $this->cache->set($cacheKey, $defn);
     }
     return $defn;
   }