Overview
----------------------------------------
The AFForm extension slows a site down. This is caused by the AFForm extension loading all meta data files.

The list of meta data files is cached by the contents it is not.

My proposal would be to cache also the content of the meta data file.

In my local environment with (a copy a large production database) the view contact timings went down from 4.36 seconds to 2.88 seconds when caching of the contents of the meta data file where enabled.

On a production site the timings of the view contact summary page went down from 7.39 seconds to 5.70 seconds.

See https://lab.civicrm.org/dev/core/-/issues/5237