@colemanw 2024-05-27
This seems like a good change indeed if it makes a big performance improvement. However the patch needs a little work because it's currently ignoring the `$getLayout` param. The layout will need to be appended, or alternately, the layout-fetching part could be moved to a separate function.