@colemanw 2024-05-28
@jaapjansma I see that you've changed it to cache the layout as well. My preference would be to not do that, as layouts can be quite large, and generally only need to be read once so I don't think there'd be much benefit to caching, at least not enough benefit to justify the extra cache storage required.
I'm leaning toward adding a new function to get the layout, and refactoring callers of this function to only call for the metadata & call the 2nd function if they need the layout.