@jaapjansma 2024-05-31
I am going to close this PR as the caching improvement introduces issues with the tests. Also my test locally and on production where wrong I had debugging enabled and after turning debugging off that was the huge performance improvemenet